start restructuring knot insertion and removal tests.
