+++ test/spl/knot_insertion_test.cc
@@ -145,162 +145,6 @@ TEST_F(NURBSEx5_2, InsertsKnot2_0Correctly) {  // NOLINT
   }
 }
 
-class Random2DBSplineForKnotInsertion : public Test {  // NOLINT
- public:
-  Random2DBSplineForKnotInsertion() {
-    std::array<ParamCoord, 2> limits = {ParamCoord{0.0}, ParamCoord{1.0}};
-    spl::RandomBSplineGenerator<2> spline_generator(limits, 10, 3);
-    spl::BSpline<2> b_spline(spline_generator);
-    bspline_2d_before_ = std::make_shared<spl::BSpline<2>>(b_spline);
-    spl::BSpline<2> b_spline_after(b_spline);
-    bspline_2d_after_ = std::make_shared<spl::BSpline<2>>(b_spline_after);
-  }
-
- protected:
-  std::shared_ptr<spl::BSpline<2>> bspline_2d_before_;
-  std::shared_ptr<spl::BSpline<2>> bspline_2d_after_;
-};
-
-TEST_F(Random2DBSplineForKnotInsertion, InsertsKnot0_4InFirstDirectionCorrectly) {  // NOLINT
-  bspline_2d_after_->InsertKnot(ParamCoord(0.4), 0);
-  ASSERT_THAT(bspline_2d_after_->GetKnotVector(0)->GetNumberOfKnots(),
-              bspline_2d_before_->GetKnotVector(0)->GetNumberOfKnots() + 1);
-  ASSERT_THAT(bspline_2d_after_->GetKnotVector(1)->GetNumberOfKnots(),
-              bspline_2d_before_->GetKnotVector(1)->GetNumberOfKnots());
-  ASSERT_THAT(bspline_2d_after_->GetPointsPerDirection()[0], bspline_2d_before_->GetPointsPerDirection()[0] + 1);
-  double steps = 50;
-  for (int i = 0; i <= steps; ++i) {
-    ParamCoord coord2 = ParamCoord{util::Random::GetUniformRandom<double>(0.0, 1.0)};
-    std::array<ParamCoord, 2> param_coord{ParamCoord(i / steps), coord2};
-    for (int j = 0; j < 2; ++j) {
-      ASSERT_THAT(bspline_2d_after_->Evaluate(param_coord, {j})[0],
-                  DoubleNear(bspline_2d_before_->Evaluate(param_coord, {j})[0], 0.000001));
-    }
-  }
-}
-
-TEST_F(Random2DBSplineForKnotInsertion, InsertsKnot0_7InSecondDirectionCorrectly) {  // NOLINT
-  bspline_2d_after_->InsertKnot(ParamCoord(0.7), 1);
-  ASSERT_THAT(bspline_2d_after_->GetKnotVector(0)->GetNumberOfKnots(),
-              bspline_2d_before_->GetKnotVector(0)->GetNumberOfKnots());
-  ASSERT_THAT(bspline_2d_after_->GetKnotVector(1)->GetNumberOfKnots(),
-              bspline_2d_before_->GetKnotVector(1)->GetNumberOfKnots() + 1);
-  ASSERT_THAT(bspline_2d_after_->GetPointsPerDirection()[1], bspline_2d_before_->GetPointsPerDirection()[1] + 1);
-  double steps = 50;
-  for (int i = 0; i <= steps; ++i) {
-    ParamCoord coord1 = ParamCoord{util::Random::GetUniformRandom<double>(0.0, 1.0)};
-    std::array<ParamCoord, 2> param_coord{coord1, ParamCoord(i / steps)};
-    for (int j = 0; j < 2; ++j) {
-      ASSERT_THAT(bspline_2d_after_->Evaluate(param_coord, {j})[0],
-                  DoubleNear(bspline_2d_before_->Evaluate(param_coord, {j})[0], 0.000001));
-    }
-  }
-}
-
-TEST_F(Random2DBSplineForKnotInsertion, InsertsKnot0_4InFirstAndKnot0_7InSecondDirectionCorrectly) {  // NOLINT
-  bspline_2d_after_->InsertKnot(ParamCoord(0.4), 0);
-  bspline_2d_after_->InsertKnot(ParamCoord(0.7), 1);
-  ASSERT_THAT(bspline_2d_after_->GetKnotVector(0)->GetNumberOfKnots(),
-              bspline_2d_before_->GetKnotVector(0)->GetNumberOfKnots() + 1);
-  ASSERT_THAT(bspline_2d_after_->GetKnotVector(1)->GetNumberOfKnots(),
-              bspline_2d_before_->GetKnotVector(1)->GetNumberOfKnots() + 1);
-  ASSERT_THAT(bspline_2d_after_->GetPointsPerDirection()[0], bspline_2d_before_->GetPointsPerDirection()[0] + 1);
-  ASSERT_THAT(bspline_2d_after_->GetPointsPerDirection()[1], bspline_2d_before_->GetPointsPerDirection()[1] + 1);
-  double steps = 10;
-  for (int i = 0; i <= steps; ++i) {
-    for (int j = 0; j <= steps; ++j) {
-      std::array<ParamCoord, 2> param_coord{ParamCoord(i / steps), ParamCoord(j / steps)};
-      for (int k = 0; k < 2; ++k) {
-        ASSERT_THAT(bspline_2d_after_->Evaluate(param_coord, {k})[0],
-                    DoubleNear(bspline_2d_before_->Evaluate(param_coord, {k})[0], 0.000001));
-      }
-    }
-  }
-}
-
-class Random3DBSplineForKnotInsertion : public Test {  // NOLINT
- public:
-  Random3DBSplineForKnotInsertion() {
-    std::array<ParamCoord, 2> limits = {ParamCoord{0.0}, ParamCoord{1.0}};
-    spl::RandomBSplineGenerator<3> spline_generator(limits, 10, 4);
-    spl::BSpline<3> b_spline(spline_generator);
-    bspline_3d_before_ = std::make_shared<spl::BSpline<3>>(b_spline);
-    spl::BSpline<3> b_spline_after(b_spline);
-    bspline_3d_after_ = std::make_shared<spl::BSpline<3>>(b_spline_after);
-  }
-
- protected:
-  std::shared_ptr<spl::BSpline<3>> bspline_3d_before_;
-  std::shared_ptr<spl::BSpline<3>> bspline_3d_after_;
-};
-
-TEST_F(Random3DBSplineForKnotInsertion, InsertsKnot0_4InFirst_Knot0_99InSecond_Knot0_01InThirdDirection) {  // NOLINT
-  bspline_3d_after_->InsertKnot(ParamCoord(0.4), 0);
-  bspline_3d_after_->InsertKnot(ParamCoord(0.99), 1);
-  bspline_3d_after_->InsertKnot(ParamCoord(0.01), 2);
-  ASSERT_THAT(bspline_3d_after_->GetKnotVector(0)->GetNumberOfKnots(),
-              bspline_3d_before_->GetKnotVector(0)->GetNumberOfKnots() + 1);
-  ASSERT_THAT(bspline_3d_after_->GetKnotVector(1)->GetNumberOfKnots(),
-              bspline_3d_before_->GetKnotVector(1)->GetNumberOfKnots() + 1);
-  ASSERT_THAT(bspline_3d_after_->GetKnotVector(2)->GetNumberOfKnots(),
-              bspline_3d_before_->GetKnotVector(2)->GetNumberOfKnots() + 1);
-  ASSERT_THAT(bspline_3d_after_->GetPointsPerDirection()[0], bspline_3d_before_->GetPointsPerDirection()[0] + 1);
-  ASSERT_THAT(bspline_3d_after_->GetPointsPerDirection()[1], bspline_3d_before_->GetPointsPerDirection()[1] + 1);
-  ASSERT_THAT(bspline_3d_after_->GetPointsPerDirection()[2], bspline_3d_before_->GetPointsPerDirection()[2] + 1);
-  double s = 5;
-  for (int i = 0; i <= s; ++i) {
-    for (int j = 0; j <= s; ++j) {
-      for (int k = 0; k <= s; ++k) {
-        std::array<ParamCoord, 3> param_coord{ParamCoord(i / s), ParamCoord(j / s), ParamCoord(k / s)};
-        for (int l = 0; l < 3; ++l) {
-          ASSERT_THAT(bspline_3d_after_->Evaluate(param_coord, {l})[0],
-                      DoubleNear(bspline_3d_before_->Evaluate(param_coord, {l})[0], 0.000001));
-        }
-      }
-    }
-  }
-}
-
-class Random3DNURBSForKnotInsertion : public Test {  // NOLINT
- public:
-  Random3DNURBSForKnotInsertion() {
-    std::array<ParamCoord, 2> limits = {ParamCoord{0.0}, ParamCoord{1.0}};
-    spl::RandomNURBSGenerator<3> spline_generator(limits, 10, 4);
-    spl::NURBS<3> nurbs(spline_generator);
-    nurbs_3d_before_ = std::make_shared<spl::NURBS<3>>(nurbs);
-    spl::NURBS<3> nurbs_after(nurbs);
-    nurbs_3d_after_ = std::make_shared<spl::NURBS<3>>(nurbs_after);
-  }
-
- protected:
-  std::shared_ptr<spl::NURBS<3>> nurbs_3d_before_;
-  std::shared_ptr<spl::NURBS<3>> nurbs_3d_after_;
-};
-
-TEST_F(Random3DNURBSForKnotInsertion, InsertsKnot0_4InFirst_Knot0_99InSecond_Knot0_01InThirdDirection) {  // NOLINT
-  nurbs_3d_after_->InsertKnot(ParamCoord(0.4), 0);
-  nurbs_3d_after_->InsertKnot(ParamCoord(0.99), 1);
-  nurbs_3d_after_->InsertKnot(ParamCoord(0.01), 2);
-  ASSERT_THAT(nurbs_3d_after_->GetKnotVector(0)->GetNumberOfKnots(),
-              nurbs_3d_before_->GetKnotVector(0)->GetNumberOfKnots() + 1);
-  ASSERT_THAT(nurbs_3d_after_->GetKnotVector(1)->GetNumberOfKnots(),
-              nurbs_3d_before_->GetKnotVector(1)->GetNumberOfKnots() + 1);
-  ASSERT_THAT(nurbs_3d_after_->GetKnotVector(2)->GetNumberOfKnots(),
-              nurbs_3d_before_->GetKnotVector(2)->GetNumberOfKnots() + 1);
-  ASSERT_THAT(nurbs_3d_after_->GetWeights().size(), nurbs_3d_after_->GetNumberOfControlPoints());
-  ASSERT_THAT(nurbs_3d_after_->GetPointsPerDirection()[0], nurbs_3d_before_->GetPointsPerDirection()[0] + 1);
-  ASSERT_THAT(nurbs_3d_after_->GetPointsPerDirection()[1], nurbs_3d_before_->GetPointsPerDirection()[1] + 1);
-  ASSERT_THAT(nurbs_3d_after_->GetPointsPerDirection()[2], nurbs_3d_before_->GetPointsPerDirection()[2] + 1);
-  std::array<ParamCoord, 3> param_coord{};
-  for (int i = 0; i < 3; ++i) {
-    param_coord[i] = ParamCoord{util::Random::GetUniformRandom<double>(0.0, 1.0)};
-  }
-  for (int l = 0; l < 3; ++l) {
-    ASSERT_THAT(nurbs_3d_after_->Evaluate(param_coord, {l})[0],
-                DoubleNear(nurbs_3d_before_->Evaluate(param_coord, {l})[0], 0.000001));
-  }
-}
-
 class BSplineFig5_16 : public Test {  // NOLINT
  public:
   BSplineFig5_16() {
